Split popover route helpers
